requirements.txt
google-generativeai
SpeechRecognition
pyaudio
